Quote Ranks
Create an app which a user can use the form to create a quote with the author's name and list the quotes under the form as shown in the wireframe, if you complete this on time, move on to complete the bonus features: 

BONUS: Each quote should include a rating value which by default starts at 0. Allow the user to increase and decrease ratings for each quote.